# openapi.yaml
...
openapi: "3.1.0"
info:
  title: FaunaLab
  version: "0.1.0"
  summary: 犬種分類システムの HTTP 契約
  description: |
    プログラムインタフェース（REQ-API-001）と観測インタフェース（REQ-OBS-001）の契約。
    設計の根拠と不採用案は [DESIGN.md](DESIGN.md) 3〜4 節。

    ## パス規約

    - リソースは名詞の複数形。プレフィックスは `/api/`。
    - 観測だけ仕様どおり `GET /api/_state`。
    - JSON フィールドは snake_case。日時は ISO 8601-1 の UTC（末尾 `Z`）。
    - 公開識別子 `ref` は UUID version 4（小文字・ハイフン付き）。削除後は再利用しない。

    ## ページネーション（REQ-API-006）

    一覧は `limit`（既定 50、上限 200）と `offset`（既定 0）を受け付け、`items` と `total` を返す。
    同一条件の順序は常に一意。ソートキーは各操作の説明と DESIGN.md 8.2 / I-API-003。

    ## 誤り（REQ-API-003 / 004 / 005）

    応答は RFC 9457 Problem Details（`application/problem+json`）。
    `code` が機械判別用識別子、`detail` が利用者向け日本語説明。
    `type` は `urn:faunalab:error:{code}`。実行時に解決しに行かない。
    スタックトレース、内部ファイルパス、SQL は含めない。

    REQ-API-003 の 7 状況は **それぞれ異なるステータス** にする（同一番号に 2 類を載せない）。

    | 状況 | HTTP | `code` の型 |
    | --- | --- | --- |
    | 要求内容の誤り | 400 | `validation_error` |
    | 対象の不在 | 404 | `{resource}_not_found` |
    | 現在の状態では実行できない操作 | 409 | 状況別（`image_duplicate` 等） |
    | 前提条件の不成立 | 422 | 状況別（`training_precondition` 等） |
    | 許可されない媒体型 | 415 | `unsupported_media_type` |
    | 上限を超える大きさ | 413 | `payload_too_large` |
    | システム内部の予期しない誤り | 500 | `internal_error` |

    語彙の列挙は `components.schemas.ErrorCode`。

    複数ファイルアップロードの件別失敗は操作全体を誤りにしない（REQ-F-IMG-005）。
    HTTP 200 と件別 `ok` / `code`。件別 `code` も同じ語彙。

    ## CORS（REQ-API-007）

    許可オリジンは環境変数 `FAUNALAB_CORS_ORIGINS`（カンマ区切り）。
...

...
  contact:
...
servers:
...
paths:
...
  /api/images:
    get:
      tags: [images]
      operationId: listImages
      summary: 画像一覧
      description: |
        絞り込みは `split`・`class_id`・`labeled` の論理積（REQ-F-IMG-008）。
        順序は `created_at` 昇順、同刻は `ref` 昇順。
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: split
          in: query
          schema:
            $ref: "#/components/schemas/Split"
        - name: class_id
          in: query
          schema:
            $ref: "#/components/schemas/ClassId"
        - name: labeled
          in: query
          description: 確定ラベルの有無。未指定なら絞り込まない。
          schema:
            type: boolean
      responses:
        "200":
          description: ページ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImagePage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [images]
      operationId: uploadImages
      summary: 画像アップロード
      description: |
        マルチパート。1 回最大 50 ファイル（REQ-F-IMG-004）。
        形式は内容判定で JPEG/PNG のみ（REQ-F-IMG-001 / 002）。1 ファイル 10 MiB（REQ-F-IMG-003）。
        フィールド名は `files`。
        1 件でも成功すれば HTTP 200。件別に `ok` と失敗時 `code`。
        SHA-256 重複は件別 `image_duplicate`（409 類の語彙。HTTP は 200）。
        要求が空、または `files` 欠落は 400。
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  minItems: 1
                  maxItems: 50
                  items:
                    type: string
                    format: binary
      responses:
        "200":
          description: 件別結果（部分成功を含む）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/images/{ref}:
    parameters:
      - $ref: "#/components/parameters/ImageRef"
    get:
      tags: [images]
      operationId: getImage
      summary: 画像詳細
      responses:
        "200":
          description: 詳細
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Image"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [images]
      operationId: deleteImage
      summary: 画像削除
      description: |
        確定ラベル・候補・推論履歴・画像ファイル・サムネイルを削除する（REQ-F-IMG-009）。
        学習済モデル版は残す。
      responses:
        "204":
          description: 削除した
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/images/{ref}/thumbnail:
    parameters:
      - $ref: "#/components/parameters/ImageRef"
    get:
      tags: [images]
      operationId: getImageThumbnail
      summary: サムネイル
      description: 長辺 256 以下、アスペクト比保持（REQ-F-IMG-007）。媒体型は JPEG。
      responses:
        "200":
          description: サムネイル画像
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  parameters:
    Limit:
      name: limit
      in: query
      description: 件数制限。既定 50、上限 200。
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Offset:
      name: offset
      in: query
      description: 先頭からの位置。0 始まり。
      schema:
        type: integer
        minimum: 0
        default: 0
    ImageRef:
      name: ref
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Ref"
...
  responses:
    BadRequest:
      description: 要求内容の誤り（REQ-API-003）
      x-srs-situation: 要求内容の誤り
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemBadRequest"
          examples:
            validation_error:
              value:
                type: urn:faunalab:error:validation_error
                title: Bad Request
                status: 400
                detail: 要求の内容が正しくありません。
                code: validation_error
    NotFound:
      description: 対象の不在（REQ-API-003）
      x-srs-situation: 対象の不在
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemNotFound"
          examples:
            image_not_found:
              value:
                type: urn:faunalab:error:image_not_found
                title: Not Found
                status: 404
                detail: 指定した画像は存在しません。
                code: image_not_found
...
    UnsupportedMediaType:
      description: 許可されない媒体型（REQ-API-003）
      x-srs-situation: 許可されない媒体型
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemUnsupportedMediaType"
          examples:
            unsupported_media_type:
              value:
                type: urn:faunalab:error:unsupported_media_type
                title: Unsupported Media Type
                status: 415
                detail: JPEG または PNG 以外の画像は受け付けません。
                code: unsupported_media_type
    PayloadTooLarge:
      description: 上限を超える大きさ（REQ-API-003）
      x-srs-situation: 上限を超える大きさ
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemPayloadTooLarge"
          examples:
            payload_too_large:
              value:
                type: urn:faunalab:error:payload_too_large
                title: Content Too Large
                status: 413
                detail: ファイルサイズが 10 MiB を超えています。
                code: payload_too_large
    InternalError:
      description: システム内部の予期しない誤り（REQ-API-003）
      x-srs-situation: システム内部の予期しない誤り
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemInternalError"
          examples:
            internal_error:
              value:
                type: urn:faunalab:error:internal_error
                title: Internal Server Error
                status: 500
                detail: 処理中に予期しない誤りが起きました。
                code: internal_error
  schemas:
    Ref:
      type: string
      format: uuid
      description: UUID version 4。対象が存在する限り不変。削除後は再利用しない。
      examples: ["aaaaaaaa-bbbb-4ccc-8ddd-eeeeeeeeeeee"]
    DateTime:
      type: string
      format: date-time
      description: UTC、末尾 Z。
      examples: ["2026-09-12T03:04:05Z"]
    ClassId:
      type: string
      enum:
        - samoyed
        - great_pyrenees
        - boxer
        - american_bulldog
        - chihuahua
        - miniature_pinscher
        - pomeranian
        - havanese
    Split:
      type: string
      enum: [train, val, test, unassigned]
    LabelSource:
      type: string
      enum: [human, model_suggested]
...
    Sha256:
      type: string
      pattern: "^[0-9a-f]{64}$"
      description: 小文字 hex。
    ErrorCode:
      type: string
      description: 機械判別用の誤り識別子（REQ-API-004）。語彙を増やす場合はリソース名を接頭辞にし、7 類のいずれかに対応付ける。
      enum:
        - validation_error
        - unsupported_media_type
        - payload_too_large
        - image_not_found
        - job_not_found
        - model_not_found
        - inference_not_found
        - suggestion_not_found
        - image_duplicate
        - job_not_cancelable
        - model_not_deletable
        - training_precondition
        - no_active_model
        - empty_test_split
        - baseline_unavailable
        - sample_unavailable
        - internal_error
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, code]
      description: |
        RFC 9457。`detail` にスタックトレース、内部パス、SQL を入れない（REQ-API-005）。
        追加フィールドでトレースを足さない。
      properties:
        type:
          type: string
          format: uri
          description: "urn:faunalab:error:{code}"
          examples: ["urn:faunalab:error:image_not_found"]
        title:
          type: string
          description: ステータスの短い英語名
        status:
          type: integer
          description: HTTP ステータス。許容値はレスポンス種別ごとの Problem* スキーマで固定する。
        detail:
          type: string
          description: 利用者向けの日本語説明
        code:
          $ref: "#/components/schemas/ErrorCode"
    ProblemBadRequest:
      allOf:
        - $ref: "#/components/schemas/Problem"
        - type: object
          properties:
            status:
              const: 400
            code:
              const: validation_error
    ProblemNotFound:
      allOf:
        - $ref: "#/components/schemas/Problem"
        - type: object
          properties:
            status:
              const: 404
            code:
              enum:
                - image_not_found
                - job_not_found
                - model_not_found
                - inference_not_found
                - suggestion_not_found
...
    ProblemUnsupportedMediaType:
      allOf:
        - $ref: "#/components/schemas/Problem"
        - type: object
          properties:
            status:
              const: 415
            code:
              const: unsupported_media_type
    ProblemPayloadTooLarge:
      allOf:
        - $ref: "#/components/schemas/Problem"
        - type: object
          properties:
            status:
              const: 413
            code:
              const: payload_too_large
    ProblemInternalError:
      allOf:
        - $ref: "#/components/schemas/Problem"
        - type: object
          properties:
            status:
              const: 500
            code:
              const: internal_error
    PageMeta:
      type: object
      required: [items, total]
      properties:
        total:
          type: integer
          minimum: 0
          description: 絞り込み後の全体件数
    ImagePage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Image"
...
    ConfirmedLabel:
      type: object
      required: [class_id, source]
      properties:
        class_id:
          $ref: "#/components/schemas/ClassId"
        source:
          $ref: "#/components/schemas/LabelSource"
    SuggestionPayload:
      type: object
      required: [class_id, confidence, model_ref]
      properties:
        class_id:
          $ref: "#/components/schemas/ClassId"
        confidence:
          type: number
          minimum: 0
          maximum: 1
        model_ref:
          $ref: "#/components/schemas/Ref"
    Image:
      type: object
      required: [ref, sha256, split, label, suggestion, original_name, size_bytes, width, height, created_at, media_type]
      properties:
        ref:
          $ref: "#/components/schemas/Ref"
        sha256:
          $ref: "#/components/schemas/Sha256"
        split:
          $ref: "#/components/schemas/Split"
        label:
          oneOf:
            - $ref: "#/components/schemas/ConfirmedLabel"
            - type: "null"
        suggestion:
          oneOf:
            - $ref: "#/components/schemas/SuggestionPayload"
            - type: "null"
        original_name:
          type: string
          description: 由来するファイル名。値として保持し、記憶域パスには使わない。
        size_bytes:
          type: integer
          minimum: 1
        width:
          type: integer
          minimum: 1
        height:
          type: integer
          minimum: 1
        created_at:
          $ref: "#/components/schemas/DateTime"
        media_type:
          type: string
          enum: [image/jpeg, image/png]
    UploadItem:
      type: object
      required: [filename, ok]
      properties:
        filename:
          type: string
        ok:
          type: boolean
        ref:
          $ref: "#/components/schemas/Ref"
        code:
          $ref: "#/components/schemas/ErrorCode"
        detail:
          type: string
    UploadResult:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/UploadItem"
...